Add headless mode support and GitHub Actions E2E workflow
Enables E2E tests to run in headless mode for CI/CD pipelines using
Chrome's new headless mode which supports extensions.

**Changes:**

1. **Shared Test Configuration** (`e2e/test-config.js`)
   - Centralized Puppeteer config
   - Auto-detects CI environment (CI=true or HEADLESS=true)
   - Uses `headless: 'new'` in CI (supports extensions)
   - Uses `headless: false` locally (for debugging)

2. **Updated All E2E Tests**
   - All 5 test suites now use shared config
   - Automatically switch between headed/headless modes
   - No code changes needed for local vs CI execution

3. **GitHub Actions Workflow** (`.github/workflows/ci.yml`)
   - New job: `extension-e2e-tests`
   - Installs Chrome stable on Ubuntu
   - Runs all 10 E2E tests in headless mode
   - Uploads screenshots on failure for debugging
   - Integrated into ci-success check

**Testing:**
- ✅ All tests pass in local headed mode
- ✅ All tests pass in headless mode (HEADLESS=true)
- ✅ ~158s runtime in headless (slightly faster than headed)

**Usage:**
```bash
make test-e2e              # Local headed mode
HEADLESS=true make test-e2e  # Local headless mode
CI=true make test-e2e       # CI mode (GitHub Actions)
```

This enables continuous E2E testing on every PR and push to ensure
the smart group merging feature works correctly across all scenarios.
