+++ .github/workflows/ci.yml
@@ -15,9 +15,9 @@ permissions:
   pull-requests: read
 
 jobs:
-  # Chrome Extension Tests and Linting
+  # Chrome Extension Unit Tests and Linting
   extension-tests:
-    name: Chrome Extension Tests
+    name: Chrome Extension Unit Tests
     runs-on: ubuntu-latest
 
     steps:
@@ -116,6 +116,49 @@ jobs:
           cd chrome-extension
           # TODO: Add ESLint configuration and run: npm run lint
           echo "ESLint not configured yet - skipping"
+
+  # Chrome Extension E2E Tests
+  extension-e2e-tests:
+    name: Chrome Extension E2E Tests
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+          cache: 'npm'
+          cache-dependency-path: chrome-extension/package-lock.json
+
+      - name: Install dependencies
+        run: |
+          cd chrome-extension
+          npm ci
+
+      - name: Install Chrome (latest stable)
+        run: |
+          wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+          echo "deb http://dl.google.com/linux/chrome/deb/ stable main" | sudo tee /etc/apt/sources.list.d/google-chrome.list
+          sudo apt-get update
+          sudo apt-get install -y google-chrome-stable
+
+      - name: Run Puppeteer E2E tests (headless)
+        run: |
+          cd chrome-extension
+          CI=true npm run test:e2e
+        env:
+          CI: true
+
+      - name: Upload E2E test screenshots (on failure)
+        uses: actions/upload-artifact@v4
+        if: failure()
+        with:
+          name: e2e-screenshots
+          path: chrome-extension/e2e/screenshots/
+          retention-days: 7
 
   # Python Tests and Linting
   python-tests:
@@ -245,18 +288,20 @@ jobs:
   ci-success:
     name: CI Success
     runs-on: ubuntu-latest
-    needs: [extension-tests, python-tests, manifest-validation, build-extension]
+    needs: [extension-tests, extension-e2e-tests, python-tests, manifest-validation, build-extension]
     if: always()
 
     steps:
       - name: Check all jobs
         run: |
-          echo "Extension Tests: ${{ needs.extension-tests.result }}"
+          echo "Extension Unit Tests: ${{ needs.extension-tests.result }}"
+          echo "Extension E2E Tests: ${{ needs.extension-e2e-tests.result }}"
           echo "Python Tests: ${{ needs.python-tests.result }}"
           echo "Manifest Validation: ${{ needs.manifest-validation.result }}"
           echo "Build Extension: ${{ needs.build-extension.result }}"
 
           if [ "${{ needs.extension-tests.result }}" != "success" ] || \
+             [ "${{ needs.extension-e2e-tests.result }}" != "success" ] || \
              [ "${{ needs.python-tests.result }}" != "success" ] || \
              [ "${{ needs.manifest-validation.result }}" != "success" ] || \
              [ "${{ needs.build-extension.result }}" != "success" ]; then
