+++ chrome-extension/e2e/scenario1.test.js
@@ -10,10 +10,7 @@
  */
 
 const puppeteer = require('puppeteer');
-const path = require('path');
-
-// Extension path (relative to e2e directory)
-const EXTENSION_PATH = path.resolve(__dirname, '..');
+const { getPuppeteerConfig } = require('./test-config');
 const EXTENSION_ID = 'lfgdhkjngeggkemgninmehdfohiebdnj'; // You may need to get this dynamically
 
 // Test URLs to open
@@ -45,16 +42,7 @@ describe('Scenario 1: First-Time Organization (Create New Groups)', () => {
 
   beforeAll(async () => {
     // Launch browser with extension loaded
-    browser = await puppeteer.launch({
-      headless: false, // Run in headed mode to see what's happening
-      args: [
-        `--disable-extensions-except=${EXTENSION_PATH}`,
-        `--load-extension=${EXTENSION_PATH}`,
-        '--no-sandbox',
-        '--disable-setuid-sandbox'
-      ],
-      defaultViewport: null
-    });
+    browser = await puppeteer.launch(getPuppeteerConfig());
 
     // Wait for the extension's service worker to load
     const extensionTarget = await browser.waitForTarget(
