+++ chrome-extension/e2e/scenario2.test.js
@@ -13,10 +13,7 @@
  */
 
 const puppeteer = require('puppeteer');
-const path = require('path');
-
-// Extension path (relative to e2e directory)
-const EXTENSION_PATH = path.resolve(__dirname, '..');
+const { getPuppeteerConfig } = require('./test-config');
 
 // Test URLs to open
 const INITIAL_URLS = {
@@ -56,16 +53,7 @@ describe('Scenario 2: Re-Organization (Update Existing Groups)', () => {
 
   beforeAll(async () => {
     // Launch browser with extension loaded
-    browser = await puppeteer.launch({
-      headless: false, // Run in headed mode to see what's happening
-      args: [
-        `--disable-extensions-except=${EXTENSION_PATH}`,
-        `--load-extension=${EXTENSION_PATH}`,
-        '--no-sandbox',
-        '--disable-setuid-sandbox'
-      ],
-      defaultViewport: null
-    });
+    browser = await puppeteer.launch(getPuppeteerConfig());
 
     // Wait for the extension's service worker to load
     const extensionTarget = await browser.waitForTarget(
